[YUNIKORN-3064] Refactor user_group_limit E2E test suite structure
