+++ test/e2e/user_group_limit/user_group_limit_test.go
@@ -25,7 +25,6 @@ import (
 	"net/url"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strings"
 	"time"
 
@@ -81,42 +80,16 @@ var (
 	}
 )
 
-var _ = ginkgo.BeforeSuite(func() {
-	_, filename, _, _ := runtime.Caller(0)
-	suiteName = common.GetSuiteName(filename)
-	// Initializing kubectl client
-	kClient = k8s.KubeCtl{}
-	Ω(kClient.SetClient()).To(gomega.BeNil())
-	// Initializing rest client
-	restClient = yunikorn.RClient{}
-	Ω(restClient).NotTo(gomega.BeNil())
-
-	yunikorn.EnsureYuniKornConfigsPresent()
-
-	ginkgo.By("Port-forward the scheduler pod")
-	var err = kClient.PortForwardYkSchedulerPod()
-	Ω(err).NotTo(gomega.HaveOccurred())
-})
-
-var _ = ginkgo.BeforeEach(func() {
-	dev = "dev" + common.RandSeq(5)
-	ginkgo.By("create development namespace")
-	ns, err := kClient.CreateNamespace(dev, nil)
-	gomega.Ω(err).NotTo(gomega.HaveOccurred())
-	gomega.Ω(ns.Status.Phase).To(gomega.Equal(v1.NamespaceActive))
-})
-
-var _ = ginkgo.AfterSuite(func() {
-	ginkgo.By("Check Yunikorn's health")
-	checks, err := yunikorn.GetFailedHealthChecks()
-	Ω(err).NotTo(HaveOccurred())
-	Ω(checks).To(gomega.Equal(""), checks)
-	ginkgo.By("Tearing down namespace: " + dev)
-	err = kClient.TearDownNamespace(dev)
-	Ω(err).NotTo(HaveOccurred())
-})
-
 var _ = ginkgo.Describe("UserGroupLimit", func() {
+
+	ginkgo.BeforeEach(func() {
+		dev = "dev" + common.RandSeq(5)
+		ginkgo.By("create development namespace")
+		ns, err := kClient.CreateNamespace(dev, nil)
+		gomega.Ω(err).NotTo(gomega.HaveOccurred())
+		gomega.Ω(ns.Status.Phase).To(gomega.Equal(v1.NamespaceActive))
+	})
+
 	ginkgo.It("Verify_maxresources_with_a_specific_user_limit", func() {
 		ginkgo.By("Update config")
 		// The wait wrapper still can't fully guarantee that the config in AdmissionController has been updated.
